Create README.md
Searching of an element in a given sequence can be done in two ways. Linear search or binary search.Linear search is a technique in which each element is checked in a sequence for searching operation. 
Binary search-Search a sorted array by repeatedly dividing the search interval in half. Begin with an interval covering the whole array. If the value of the search key is less than the item in the middle of the interval, narrow the interval to the lower half. Otherwise narrow it to the upper half. Repeatedly check until the value is found or the interval is empty